.ado/pipelines/python/requirements.txt
azure-cosmos==4.6.0
